Separated out routes into their own folder. Refactored controllers. Almost finished ReadMe.

+++ src/models/album.js
@@ -2,7 +2,7 @@ module.exports = (connection, DataTypes) => {
    const schema = {
       artistName: DataTypes.STRING,
       name: DataTypes.STRING,
-      year: DataTypes.INTEGER(4),
+      year: DataTypes.INTEGER,
    }
 
    const AlbumModel = connection.define('album', schema)
